Move cll function to util
